Exception Handling

+++ 03_ApiRest/DScommerce2/src/main/java/com/joaoldantasn/DScommerce/services/ProductService.java
@@ -1,15 +1,20 @@
 package com.joaoldantasn.DScommerce.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.joaoldantasn.DScommerce.dto.ProductDTO;
 import com.joaoldantasn.DScommerce.entities.Product;
+import com.joaoldantasn.DScommerce.exceptions.DataBaseException;
 import com.joaoldantasn.DScommerce.exceptions.ResourceNotFoundException;
 import com.joaoldantasn.DScommerce.repositories.ProductRepository;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
@@ -41,16 +46,28 @@ public class ProductService {
 	
 	@Transactional
 	public ProductDTO update(Long id, ProductDTO dto) {
-		//nao vai no baco de dados
-		Product entity = repository.getReferenceById(id);
-		copyDtoToEntity(dto, entity);
-		entity = repository.save(entity);
-		return new ProductDTO(entity);
+		try {
+			//nao vai no baco de dados
+			Product entity = repository.getReferenceById(id);
+			copyDtoToEntity(dto, entity);
+			entity = repository.save(entity);
+			return new ProductDTO(entity);
+		} catch(EntityNotFoundException e) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
 	}
 	
-	@Transactional
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
-		repository.deleteById(id);
+		if (!repository.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
+		try {
+	        	repository.deleteById(id);    		
+		}
+	    	catch (DataIntegrityViolationException e) {
+	        	throw new DataBaseException("Falha de integridade referencial");
+	   	}
 	}
 
 	private void copyDtoToEntity(ProductDTO dto, Product entity) {
